@@ -1,13 +1,13 @@
 {
-  "name": "next",
-  "version": "0.1.0",
+  "name": "app-yobta",
+  "version": "0.0.1",
   "private": true,
   "engines": {
     "node": ">=20.10.0"
   },
   "packageManager": "pnpm@8.15.6",
   "scripts": {
-    "dev": "next dev --turbo",
+    "dev": "next dev --turbopack",
     "build": "next build",
     "prettify": "npx prettier --wrinte .",
     "start": "next start",
@@ -18,54 +18,50 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@mdx-js/loader": "^3.0.1",
-    "@mdx-js/react": "^3.0.1",
-    "@next/mdx": "^14.2.13",
+    "@mdx-js/loader": "^3.1.0",
+    "@mdx-js/react": "^3.1.0",
+    "@next/mdx": "^15.0.3",
     "@tailwindcss/typography": "^0.5.15",
     "@yobta/stores": "0.3.2",
-    "@yobta/ui": "^0.0.51",
     "@yobta/validator": "^1.0.0",
     "@yobta/validator-react": "^0.0.2",
     "clsx": "^2.1.1",
-    "lodash": "^4.17.21",
-    "next": "^14.2.13",
+    "daisyui": "^4.12.14",
+    "next": "^15.0.3",
     "react": "18.3.1",
-    "react-dom": "18.3.1",
-    "react-use": "^17.5.1",
-    "sass": "^1.79.4",
-    "typescript-eslint": "^8.7.0",
-    "webpack": "^5.95.0"
+    "react-dom": "18.3.1"
   },
   "devDependencies": {
-    "@eslint/compat": "^1.1.1",
-    "@eslint/eslintrc": "^3.1.0",
-    "@eslint/js": "^9.11.1",
-    "@logux/eslint-config": "^53.4.0",
-    "@types/lodash": "^4.17.9",
+    "@eslint/compat": "^1.2.3",
+    "@eslint/eslintrc": "^3.2.0",
+    "@eslint/js": "^9.15.0",
+    "@logux/eslint-config": "^53.5.1",
     "@types/mdx": "^2.0.13",
-    "@types/node": "22.7.4",
-    "@types/react": "18.3.10",
-    "@types/react-dom": "18.3.0",
-    "@typescript-eslint/eslint-plugin": "^8.7.0",
-    "@typescript-eslint/parser": "^8.7.0",
-    "@vitest/coverage-v8": "^2.1.1",
+    "@types/node": "22.9.0",
+    "@types/postcss-js": "^4.0.4",
+    "@types/react": "18.3.12",
+    "@types/react-dom": "18.3.1",
+    "@typescript-eslint/eslint-plugin": "^8.15.0",
+    "@typescript-eslint/parser": "^8.15.0",
+    "@vitest/coverage-v8": "^2.1.5",
     "autoprefixer": "^10.4.20",
     "c8": "^10.1.2",
     "dotenv": "^16.4.5",
-    "eslint": "9.11.1",
-    "eslint-config-next": "14.2.13",
+    "eslint": "9.15.0",
+    "eslint-config-next": "15.0.3",
     "eslint-config-standard": "^17.1.0",
-    "eslint-plugin-import": "^2.30.0",
-    "eslint-plugin-n": "^17.10.3",
-    "eslint-plugin-perfectionist": "^3.7.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-n": "^17.13.2",
+    "eslint-plugin-perfectionist": "^3.9.1",
     "eslint-plugin-prefer-let": "^4.0.0",
     "eslint-plugin-promise": "^7.1.0",
-    "happy-dom": "^15.7.4",
-    "postcss": "^8.4.47",
+    "happy-dom": "^15.11.6",
+    "postcss": "^8.4.49",
+    "postcss-js": "^4.0.1",
     "prettier": "^3.3.3",
-    "styled-jsx-plugin-postcss": "^4.0.1",
-    "tailwindcss": "^3.4.13",
-    "typescript": "5.6.2",
-    "vitest": "^2.1.1"
+    "tailwindcss": "^3.4.15",
+    "typescript": "5.6.3",
+    "typescript-eslint": "^8.15.0",
+    "vitest": "^2.1.5"
   }
 }
