@@ -1,7 +1,6 @@
 'use client'
-import css from './layout.module.scss'
+import css from './layout.module.css'
 
-import { Home } from '@yobta/ui'
 import Link from 'next/link'
 import type { FunctionComponent, ReactNode } from 'react'
 
@@ -13,7 +12,6 @@ const MdLayout: FunctionComponent<Props> = ({ children }) => {
   return (
     <>
       <Link className="yobta-menu-item" href="/">
-        <Home />
         Home
       </Link>
       <main className={css.mdLayout}>{children}</main>
